mozilla_access_event_consumer/examples/gke/app/requirements.txt
google-cloud-pubsub==2.*
mozlog
